14. Przenisienie klasy Node do klasy IntLinkedList jako prywatna klasa zagnieżdzona oraz usunięcie pliku Node.java
